// src/app/openapi/generated-angular-sdk/model/shelfProduct.ts
/**
 * Fridge API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShelfProduct { 
    id?: number;
    name: string;
    shelf_id: number;
    quantity: number;
    status: ShelfProduct.StatusEnum;
    unit: ShelfProduct.UnitEnum;
    expiration_date: string;
}
export namespace ShelfProduct {
    export type StatusEnum = 'opened' | 'unopened';
    export const StatusEnum = {
        Opened: 'opened' as StatusEnum,
        Unopened: 'unopened' as StatusEnum
    };
    export type UnitEnum = 'g' | 'kg' | 'ml' | 'l' | 'pcs';
    export const UnitEnum = {
        G: 'g' as UnitEnum,
        Kg: 'kg' as UnitEnum,
        Ml: 'ml' as UnitEnum,
        L: 'l' as UnitEnum,
        Pcs: 'pcs' as UnitEnum
    };
}